refactor: separate out into iter and ser modules
