+++ src/note/mod.rs
@@ -140,7 +140,7 @@ impl NoteSerialized {
     pub(crate) fn new(id: String, note: Note, backlinks: Vec<String>) -> Self {
         Self {
             id,
-            links: note.links().map(|id| hex(id)).collect::<Vec<String>>(),
+            links: note.links().map(hex).collect::<Vec<String>>(),
             title: note.title,
             path: note.path,
             tags: note.tags,
